+++ R/tar_seed_create.R
@@ -4,8 +4,8 @@
 #' @description Create a seed for a target.
 #' @section Seeds:
 #'   A target's random number generator seed
-#'   is a deterministic function of its name and the global pipeline seed.
-#'   Consequently,
+#'   is a deterministic function of its name and the global pipeline seed
+#'   from [tar_option_get("seed")]. Consequently,
 #'
 #'     1. Each target runs with a reproducible seed so that
 #'        different runs of the same pipeline in the same computing
@@ -24,14 +24,14 @@
 #'   correlated results. (For a discussion of the motivating problem,
 #'   see the Section 6: "Random-number generation" in the `parallel`
 #'   package vignette: `vignette(topic = "parallel", package = "parallel")`.)
-#'   However, this risk is extremely small in practice.
-#'   
-#'   `targets` and `tarchetypes` take the approach discussed in
+#'   However, this risk is extremely small in practice, as shown by
 #'   L'Ecuyer et al. (2027) <https://doi.org/10.1016/j.matcom.2016.05.005>
-#'   "A single RNG with a 'random' seed for each stream" (Section 4:
+#'   under "A single RNG with a 'random' seed for each stream" (Section 4:
 #'   under "How to produce parallel streams and substreams").
-#'   Here, [tar_seed_create()] plays the role
-#'   of the upstream pseudo-random number generator (RNG) that produces
+#'
+#'   `targets` and `tarchetypes` take the approach discussed in the
+#'   aforementioned section of the paper, where [tar_seed_create()] plays the
+#'   role of the upstream pseudo-random number generator (RNG) that produces
 #'   seeds for the subsequent parallel streams. Specifically,
 #'   [tar_seed_create()] acts as a counter-based RNG,
 #'   where the output function is the SHA512 hash algorithm.